Gestione risposte http BlogUserController
- Gestione risposte http del BlogUserController utilizzando ResponseEntity
- deleteUserById non restituisce un body nel caso vada a buon fine (204 no content)

+++ Api-MicroBlog/src/main/java/com/bachir/apimicroblog/controller/BlogUserController.java
@@ -6,12 +6,16 @@
 package com.bachir.apimicroblog.controller;
 
 import com.bachir.apimicroblog.domain.BlogUser;
+import com.bachir.apimicroblog.entities.JsonResponseBody;
 import com.bachir.apimicroblog.service.BlogUserService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import java.util.List;
+import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -30,31 +34,72 @@ public class BlogUserController
     @Autowired
     BlogUserService blogUserService;
     
-    @RequestMapping(method = RequestMethod.GET)
+    @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation(value = "returns all the users")
-    public List<BlogUser> getUsers()
+    public ResponseEntity<JsonResponseBody> getUsers()
     {
-        return blogUserService.getAllUsers();
+        try
+        {
+            List<BlogUser> users = blogUserService.getAllUsers();
+            return ResponseEntity.status(HttpStatus.OK).body(new JsonResponseBody(HttpStatus.OK.value(), users));
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new JsonResponseBody(HttpStatus.BAD_REQUEST.value(), "Error: " + e.toString()));
+        }
     }
     
     @RequestMapping(method = RequestMethod.GET,value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation(value = "returns a user by an ID")
-    public BlogUser getUserById(@ApiParam@PathVariable("id") Long userId)
+    public ResponseEntity<JsonResponseBody> getUserById(@ApiParam@PathVariable("id") Long userId)
     {
-        return blogUserService.getBlogUserById(userId);
+        try
+        {
+            
+            BlogUser user = blogUserService.getBlogUserById(userId).get();
+            return ResponseEntity.status(HttpStatus.FOUND).body(new JsonResponseBody(HttpStatus.FOUND.value(), user));
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new JsonResponseBody(HttpStatus.NOT_FOUND.value(), "Error: " + e.toString()));
+        }
     }
     
-    @RequestMapping(method = RequestMethod.POST)
+    @RequestMapping(method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation(value = "adds a user into the database")
-    public void addUser(@ApiParam@RequestBody BlogUser user)
+    public ResponseEntity<JsonResponseBody> addUser(HttpServletRequest request,@ApiParam@RequestBody BlogUser u)
     {
-        blogUserService.insertUser(user);
+        try
+        {   
+            BlogUser user = blogUserService.insertUser(u);
+            return ResponseEntity.status(HttpStatus.CREATED).header("location", request.getRequestURL().toString() + "/" + user.getId()).body(new JsonResponseBody(HttpStatus.CREATED.value(), null));
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new JsonResponseBody(HttpStatus.BAD_REQUEST.value(), "Error: " + e.toString()));
+        }
     }
     
-    @RequestMapping(method = RequestMethod.DELETE,value = "/deleteuser/{id}")
+    @RequestMapping(method = RequestMethod.DELETE,value = "/deleteuser/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation(value = "deletes a user by an ID")
-    public void deleteUserById(@ApiParam@PathVariable("id") Long userId)
+    public ResponseEntity<JsonResponseBody> deleteUserById(@ApiParam@PathVariable("id") Long userId)
     {
-        blogUserService.deleteBlogUserById(userId);
+        try
+        {   
+            if( blogUserService.getBlogUserById(userId).isPresent())
+            {
+                blogUserService.deleteBlogUserById(userId);
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new JsonResponseBody(HttpStatus.NO_CONTENT.value(), null));
+            }
+            else
+            {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new JsonResponseBody(HttpStatus.NOT_FOUND.value(), "Error: user not found"));
+            }
+                
+        }
+        catch( Exception e )
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new JsonResponseBody(HttpStatus.BAD_REQUEST.value(), "Error: " + e.toString()));
+        }
     }
 }
